requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
google-generativeai>=0.5.2
huggingface-hub>=0.20.3
openai>=1.14.2
python-dotenv>=1.0.1
uvicorn[standard]>=0.27.1
